finished IQueryable
