+++ 06-IQueryable/IQueryable/MyQueryTranslator.cs
@@ -1,0 +1,230 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace IQueryableTask
+{
+    public class MyQueryTranslator : ExpressionVisitor
+    {
+        private StringBuilder sb;
+        private string _select = string.Empty;
+        private string _contains = string.Empty;
+        private string _whereClause = string.Empty;
+
+        public string Contains
+        {
+            get
+            {
+                return _contains;
+            }
+        }
+
+        public string Select
+        {
+            get
+            {
+                return _select;
+            }
+        }
+
+        public string WhereClause
+        {
+            get
+            {
+                return _whereClause;
+            }
+        }
+
+        public string Translate(Expression expression)
+        {
+            sb = new StringBuilder();
+            Visit(expression);
+            _whereClause = sb.ToString();
+
+            if (_whereClause.Contains("FullName") || _whereClause.Contains("LastName"))
+            {
+                throw new NotSupportedException("Not use command!");
+            }
+
+            return _whereClause;
+        }
+
+        private static Expression StripQuotes(Expression e)
+        {
+            while (e.NodeType == ExpressionType.Quote)
+            {
+                e = ((UnaryExpression)e).Operand;
+            }
+            return e;
+        }
+
+        protected override Expression VisitMethodCall(MethodCallExpression m)
+        {
+            if (m.Method.DeclaringType == typeof(Queryable) && m.Method.Name == "Where")
+            {
+                Visit(m.Arguments[0]);
+                sb.Append(" WHERE ");
+                Visit(m.Arguments[1]);
+                return m;
+            }
+            else if (m.Method.Name == "Contains")
+            {
+                Visit(m.Object);
+                sb.Append(" LIKE ");
+                Visit(m.Arguments[0]);
+                return m;
+            }
+
+            throw new InvalidOperationException(string.Format("The method '{0}' is not supported", m.Method.Name));
+        }
+
+        protected override Expression VisitUnary(UnaryExpression u)
+        {
+            switch (u.NodeType)
+            {
+                case ExpressionType.Quote:
+                    Visit(u.Operand);
+                    break;
+                case ExpressionType.Convert:
+                    Visit(u.Operand);
+                    break;
+                default:
+                    throw new NotSupportedException(string.Format("The unary operator '{0}' is not supported", u.NodeType));
+            }
+            return u;
+        }
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="b"></param>
+        /// <returns></returns>
+        protected override Expression VisitBinary(BinaryExpression b)
+        {
+            Visit(b.Left);
+
+            switch (b.NodeType)
+            {
+                case ExpressionType.And:
+                    sb.Append(" AND ");
+                    break;
+
+                case ExpressionType.AndAlso:
+                    sb.Append(" AND ");
+                    break;
+
+                case ExpressionType.Or:
+                    sb.Append(" OR ");
+                    break;
+
+                case ExpressionType.OrElse:
+                    sb.Append(" OR ");
+                    break;
+
+                case ExpressionType.Equal:
+                    if (IsNullConstant(b.Right))
+                    {
+                        sb.Append(" IS ");
+                    }
+                    else
+                    {
+                        sb.Append(" = ");
+                    }
+                    break;
+
+                case ExpressionType.NotEqual:
+                    if (IsNullConstant(b.Right))
+                    {
+                        sb.Append(" IS NOT ");
+                    }
+                    else
+                    {
+                        sb.Append(" <> ");
+                    }
+                    break;
+
+                case ExpressionType.LessThan:
+                    sb.Append(" < ");
+                    break;
+
+                case ExpressionType.LessThanOrEqual:
+                    sb.Append(" <= ");
+                    break;
+
+                case ExpressionType.GreaterThan:
+                    sb.Append(" > ");
+                    break;
+
+                case ExpressionType.GreaterThanOrEqual:
+                    sb.Append(" >= ");
+                    break;
+
+                default:
+                    throw new NotSupportedException(string.Format("The binary operator '{0}' is not supported", b.NodeType));
+            }
+
+            Visit(b.Right);
+            return b;
+        }
+
+        protected override Expression VisitConstant(ConstantExpression c)
+        {
+            IQueryable q = c.Value as IQueryable;
+
+            if (q != null)
+            {
+                sb.Append("SELECT * FROM person");
+            }
+            else
+            {
+                switch (Type.GetTypeCode(c.Value.GetType()))
+                {
+                    case TypeCode.Boolean:
+                        sb.Append(((bool)c.Value) ? 1 : 0);
+                        break;
+
+                    case TypeCode.String:
+                        sb.Append("'%");
+                        sb.Append(c.Value);
+                        sb.Append("%'");
+                        break;
+
+                    case TypeCode.DateTime:
+                        sb.Append("'%");
+                        sb.Append(c.Value);
+                        sb.Append("%'");
+                        break;
+
+                    case TypeCode.Object:
+                        throw new NotSupportedException(string.Format("The constant for '{0}' is not supported", c.Value));
+
+                    default:
+                        sb.Append(c.Value);
+                        break;
+                }
+            }
+
+            return c;
+        }
+
+        protected override Expression VisitMember(MemberExpression m)
+        {
+            if (m.Expression != null && m.Expression.NodeType == ExpressionType.Parameter)
+            {
+                sb.Append(m.Member.Name);
+                return m;
+            }
+
+            throw new NotSupportedException(string.Format("The member '{0}' is not supported", m.Member.Name));
+        }
+
+        protected bool IsNullConstant(Expression exp)
+        {
+            return (exp.NodeType == ExpressionType.Constant && ((ConstantExpression)exp).Value == null);
+        }
+    }
+
+}
