+++ 06-IQueryable/IQueryable/People.cs
@@ -38,14 +38,12 @@ namespace IQueryableTask
         public IEnumerator<Person> GetEnumerator()
         {
             // TODO: Implement GetEnumerator
-            //throw new NotImplementedException();
             return (Provider.Execute<IEnumerable<Person>>(Expression)).GetEnumerator();
         }
 
         IEnumerator IEnumerable.GetEnumerator()
         {
             // TODO: Implement GetEnumerator
-            //throw new NotImplementedException();
             return (Provider.Execute<System.Collections.IEnumerable>(Expression)).GetEnumerator();
         }
 
@@ -56,7 +54,6 @@ namespace IQueryableTask
             get
             {
                 // TODO: Implement GetEnumerator
-                //throw new NotImplementedException();
                 return typeof(Person);
             }
         }
